requirements.txt
nltk==3.9.1
openai==1.61.0
pydub==0.25.1
python-dotenv==1.0.1
